requirements.txt
streamlit>=1.28.0
selenium>=4.15.0
webdriver-manager>=4.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
beautifulsoup4>=4.12.0
reportlab>=4.0.0